Update alabama_SignUp_01_Router.js
Sarai Hannah Ajai have written a JavaScript coded constant flash equal to the "require('express-flash'); import the express-flash library/module, which provides a middleware for storing and displaying flash messages in an "alabama_SignUp_01.ejs" web page. And, these flash messages are short-lived messages that are stored in the session and displayed to an users on the "next" requests. Furthermore, Ms. Ajai can easily create and manage flash messages into the iVoteBallot web application for which can be used to display success, error or any other kind of notifications to the front-end users on to the device screens dated on 11-04-23.