Create alabamaVoters_SignUpLogin_01.ejs
The Developer, Sarai Hannah Ajai have created and written her HTML and JavaScript codes languages for the purpose of the "alabamaVoters_SignUpLogin_01.ejs" file to which provide an interface for front-end users to sign up for an account onto the iVoteBallot's website or web application specifically for the state of Alabama. The "alabamaVoters_SignUpLogin_01.ejs" web page includes a sign-up form with various 	input fields and options for users to provide their information.

The "alabamaVoters_SignUpLogin_01.ejs" web page starts with a header section that includes common header elements shared across multiple pages. The main content of the page is contained within a flex container, which is divided into two columns. The first column contains the sign-up functionality, while the second column contains the login functionality.

In the first column, there is a "Sign Up | Alabama" button that triggers a pop-up sign-up
menu when clicked. The sign-up menu includes a form where users can enter their Alabama 	AlphaNumeric card codes, Alabama identification card information, email address, confirm email address, and phone number. The sign up form also includes validation checks for the users entered data information.

The second column contains a "Log In | Alabama" button that triggers a pop-up login menu when clicked. The login menu includes a form where users can enter their email address and password in order to log in to their account. And, both forms also includes a "Forgot Password?" link for password recovery and a link to the sign-up page for new users.

Overall, the "alabamaVoter_SignUpLogin_01.ejs" web page serves, as an entry point for the users in the state of Alabama to sign up for an account on the iVoteBallot website or web application, providing an user-friendly interface and ensuring data validations for the user entered data information dated on 11-20-23 at approximately 05:37pm.