Update index.js
Sarai Hannah Ajai has opted to relocate the router and controller responsible for the passport, session, bcrypt, and SQLite3 libraries to the "index.js" (server). This adjustment was made due to issues within the MVC, specifically in properly reading the user during deserialization to the session cookie. The aim is to facilitate the retrieval of an user's data information in response to his/her request dated December 22, 2023, around 07:55PM.